feat: 🎸 refactor handlers
