+++ src/mocks/handlers.ts
@@ -1,3 +1,4 @@
-import { deleteClassCode, getClassCode, getSubjects } from "./subjectsHandlers";
+import { getClassCode, deleteClassCode } from "./classHandlers";
+import { getSubjects } from "./subjectsHandlers";
 
 export const handlers = [getSubjects, getClassCode, deleteClassCode];
