+++ src/mocks/subjectsHandlers.ts
@@ -21,17 +21,3 @@ export const getSubjects = rest.get("/subjects", (req, res, ctx) => {
     ])
   );
 });
-
-// Handle a GET for class codes
-export const getClassCode = rest.get("/class-code/:id", (req, res, ctx) => {
-  return res(
-    ctx.status(200),
-    ctx.json({code: "JFASSMDA"})
-  );
-});
-
-export const deleteClassCode = rest.delete("/class-code/:id", (req, res, ctx) => {
-  return res(
-    ctx.status(200)
-  )
-});
